+++ purrcafe/_database/_users.py
@@ -4,7 +4,7 @@ from typing import Final
 import typing
 
 from ._database import database as db, database_lock as db_l, _Nothing
-from .exceptions import WrongHashLengthError, IDNotFoundError, ObjectIDUnknownError, WrongValueLengthError, ValueMismatchError, ObjectNotFound
+from .exceptions import WrongHashLengthError, IDNotFoundError, ObjectIDUnknownError, WrongValueLengthError, ValueMismatchError, ObjectNotFound, OperationPermissionError
 from ..meowid import MeowID
 if typing.TYPE_CHECKING:
     from ._sessions import Session
@@ -38,6 +38,9 @@ class User:
 
     @name.setter
     def name(self, new_name: str) -> None:
+        if int(self.id) == 0:
+            raise OperationPermissionError("changing guest user properties")
+
         if len(new_name) > self.NAME_MAX_LENGTH:
             raise WrongValueLengthError("name", "characters", self.NAME_MAX_LENGTH, None, len(new_name))
 
@@ -57,6 +60,9 @@ class User:
 
     @email.setter
     def email(self, new_email: str) -> None:
+        if int(self.id) == 0:
+            raise OperationPermissionError("changing guest user properties")
+
         with db_l.writer:
             db.execute("UPDATE users SET email=(?) WHERE id=(?)", (new_email, int(self.id)))
             db.commit()
@@ -73,6 +79,9 @@ class User:
 
     @password_hash.setter
     def password_hash(self, new_password_hash: str) -> None:
+        if int(self.id) == 0:
+            raise OperationPermissionError("changing guest user properties")
+
         if len(new_password_hash) != self.PASSWORD_HASH_LENGTH:
             raise WrongHashLengthError("password", self.PASSWORD_HASH_LENGTH, len(new_password_hash))
 
@@ -140,6 +149,9 @@ class User:
         return user
 
     def authorize(self, password_hash: str, lifetime: datetime.timedelta = datetime.timedelta(days=7)) -> Session:  # i LOVE circular dependency error
+        if int(self.id) == 0:
+            raise OperationPermissionError("logging into guest account")
+
         from ._sessions import Session
 
         if self.password_hash != password_hash:
@@ -173,6 +185,9 @@ class User:
         return cls(*raw_data)
 
     def delete(self) -> None:
+        if int(self.id) == 0:
+            raise OperationPermissionError("deletion of guest user")
+
         for session in self.sessions:
             session.delete()
